Implement non-JavaScript functionality, more code styling.
This commit correctly implements the quiz checking for users who have
JavaScript disabled. Therefore, it also removes the "Please enable
JavaScript" message, as JavaScript is no longer needed to view or answer
the quiz questions.

If a user does not answer the quiz correctly, the comment is not posted.
Instead, the user is redirected back to the original comment form with
their comment content intact and is told to try the quiz again. This
implementation simply uses query string parameters for this to avoid
additional load on the database or storing any persistent data.

The `rawurlencode()` and `rawurldecode()` along with the `esc_html()`
functions are carefully used to avoid reflected XSS attacks.

Additionally, more pieces of the codebase have been modified to use the
WP coding style. See https://make.wordpress.org/core/handbook/best-practices/coding-standards/php/
for more information on the PHP coding style guidelines.