Remove dependency on JavaScript for quiz editing.
This commit makes it possible to add an arbitrary number of quiz
questions to a post's quiz form without the need for JavaScript on the
admin side. This is done by padding the quiz array however many
questions are in the current quiz, plus one. The JavaScript button used
to add another question is hidden if JavaScript is not available, since
that button does nothing in that case.

Also, slightly more semantic HTML is used. The `<br/>` elements have
been removed. The questions are now marked up using `<p>` and `<li>`s.

The commit also extracts the PHP and HTML from the JavaScript into its
own function. The code is also restyled to match the WP coding style.