Modify the JavaScript component so it works seamlessly with no-JS WP.
With this commit, the JavaScript component now renders itself inside a
`FORM` element, rather than outside. This begins addressing the issues
listed in #1, and means that both JavaScript and non-JavaScript
variations work; the JavaScript now creates inputs as the WordPress
back-end expects them.

Further, this commit also fixes the error animation, and changes the
animation for when the user enters the quiz answers correctly.
Previously, this component hid the comment form and revealed it when the
quiz was answered, but this required rendering the quiz outside the
form. With the above change, the animation has been changed to disabling
the comment form's textareas and re-enabling them when the quiz is
answered correctly, along with showing a bolded green success message.